ai-engine/requirements.txt
opencv-python-headless==4.8.1.78
deepface==0.0.89
openai-whisper==20231117
sentence-transformers==2.2.2
scikit-learn
numpy
torch
